chore(context): decouple context and touchable
