+++ src/app/Components/ContextMenu/ContextMenuArtwork.tsx
@@ -1,4 +1,3 @@
-import { useColor } from "@artsy/palette-mobile"
 import { ArtworkGridItem_artwork$data } from "__generated__/ArtworkGridItem_artwork.graphql"
 import { ArtworkRailCard_artwork$data } from "__generated__/ArtworkRailCard_artwork.graphql"
 import { useShareSheet } from "app/Components/ShareSheet/ShareSheetContext"
@@ -8,12 +7,7 @@ import { useFeatureFlag } from "app/utils/hooks/useFeatureFlag"
 import { useSaveArtwork } from "app/utils/mutations/useSaveArtwork"
 import { Schema } from "app/utils/track"
 import React from "react"
-import {
-  InteractionManager,
-  Platform,
-  TouchableHighlight,
-  TouchableHighlightProps,
-} from "react-native"
+import { InteractionManager, Platform } from "react-native"
 import ContextMenu, { ContextMenuAction, ContextMenuProps } from "react-native-context-menu-view"
 import Haptic, { HapticFeedbackTypes } from "react-native-haptic-feedback"
 import { useTracking } from "react-tracking"
@@ -22,29 +16,18 @@ interface ContextAction extends Omit<ContextMenuAction, "subtitle"> {
   onPress?: () => void
 }
 
-interface ContextMenuArtworkProps extends TouchableHighlightProps {
+interface ContextMenuArtworkProps {
   artwork: ArtworkRailCard_artwork$data | ArtworkGridItem_artwork$data
   onCreateAlertActionPress: () => void
-
-  /**
-   * `haptic` can be used like:
-   *  `<ContextMenuArtwork haptic />`
-   * or
-   * `<ContextMenuArtwork haptic="impactHeavy" />`
-   */
-  haptic?: HapticFeedbackTypes | true
+  haptic?: HapticFeedbackTypes | boolean
 }
 
 export const ContextMenuArtwork: React.FC<ContextMenuArtworkProps> = ({
   artwork,
   children,
-  haptic,
-  onPress,
-  onLongPress,
+  haptic = true,
   onCreateAlertActionPress,
-  ...touchableProps
 }) => {
-  const color = useColor()
   const { trackEvent } = useTracking()
   const { showShareSheet } = useShareSheet()
   const enableInstantVIR = useFeatureFlag("AREnableInstantViewInRoom")
@@ -73,7 +56,6 @@ export const ContextMenuArtwork: React.FC<ContextMenuArtworkProps> = ({
     trackEvent({
       action_name: Schema.ActionNames.ViewInRoom,
       action_type: Schema.ActionTypes.Tap,
-      // TODO: do we need tracking here?
       context_module: Schema.ContextModules.ArtworkActions,
     })
 
@@ -144,26 +126,13 @@ export const ContextMenuArtwork: React.FC<ContextMenuArtworkProps> = ({
 
   const contextActions = getContextMenuActions()
 
-  const triggerHaptic = () => {
+  const handleContextPress: ContextMenuProps["onPress"] = (event) => {
     if (haptic) {
       Haptic.trigger(haptic === true ? "impactLight" : haptic)
     }
-  }
 
-  const handleTouchablePress: TouchableHighlightProps["onPress"] = (event) => {
-    triggerHaptic()
-    onPress?.(event)
-  }
-
-  const handleTouchableLongPress: TouchableHighlightProps["onLongPress"] = (event) => {
-    triggerHaptic()
-    onLongPress?.(event)
-  }
-
-  const handleContextPress: ContextMenuProps["onPress"] = (event) => {
     const onPressToCall = contextActions[event.nativeEvent.index].onPress
 
-    triggerHaptic()
     onPressToCall?.()
   }
 
@@ -173,15 +142,7 @@ export const ContextMenuArtwork: React.FC<ContextMenuArtworkProps> = ({
       onPress={handleContextPress}
       disabled={!shouldDisplayContextMenu}
     >
-      <TouchableHighlight
-        underlayColor={color("white100")}
-        activeOpacity={0.8}
-        {...touchableProps}
-        onPress={handleTouchablePress}
-        onLongPress={handleTouchableLongPress}
-      >
-        {children}
-      </TouchableHighlight>
+      {children}
     </ContextMenu>
   )
 }
